make sure connections are addressed properly in controllers
As we have multiple "connection" objects in one widget 'this' always
refers to the latest connection object that has been created within a
given widget. In a normal method call that wouldn't cause trouble but
controllers only exist once and must work for connections. Hence we pass
the index of a connection in aCVWidget.midiOrOscconnections within the
'moreArgs' to the controller. Now 'tearDown' in unit tests works as
expected :)

Signed-off-by: Stefan Nussbaumer <st9fan@gmail.com>