hardened creating MidiConnections and connecting MIDI quite a bit
It seems to essential to isolate MIDIFuncs within a MidiConnection
against MIDIFuncs in other MidiConnections. The problem seems to be the
model/controller configuartion where all MidiConnections are kept in an
array within one model but the controller is just built around a single
function that has to detect the right MidiConnection by itself. Holding
destinct MIDIFuncs for each MidiConnection in a List in a classvar
allMidiFuncs seems to do the trick.

Signed-off-by: Stefan Nussbaumer <st9fan@gmail.com>